int arrays changed for uint8_t or uint16_t

+++ src/arduino/libraries/ServoControl/ServoControl.cpp
@@ -1,14 +1,14 @@
 #include "ServoControl.h"
 
-void ServoControl::init(int numberOfServos, int servoPins[], int servoPosMax[],
-                        int servoPosMin[], int servoPosStart[]) {
+void ServoControl::init(uint8_t numberOfServos, uint8_t servoPins[], uint16_t servoPosMax[],
+                        uint16_t servoPosMin[], uint16_t servoPosStart[]) {
     // Create vectors with the corresponding size
     this->numberOfServos = numberOfServos;
-    this->servoPins = new int[numberOfServos];
-    this->servoPosMax = new int[numberOfServos];
-    this->servoPosMin = new int[numberOfServos];
-    this->servoPosStart = new int[numberOfServos];
-    this->servoPos = new int[numberOfServos];
+    this->servoPins = new uint8_t[numberOfServos];
+    this->servoPosMax = new uint16_t[numberOfServos];
+    this->servoPosMin = new uint16_t[numberOfServos];
+    this->servoPosStart = new uint16_t[numberOfServos];
+    this->servoPos = new uint16_t[numberOfServos];
 
     // Copy the values from the parameters
     memcpy(this->servoPins, servoPins, sizeof(servoPins[0])*numberOfServos);
@@ -20,14 +20,14 @@ void ServoControl::init(int numberOfServos, int servoPins[], int servoPosMax[],
     // Iniitalize servos
     this->servos = new Servo[numberOfServos];
 
-    for (int i=0; i<numberOfServos; i++) {
+    for (uint8_t i=0; i<numberOfServos; i++) {
         servos[i].attach(servoPins[i]); // Servo setup
         servos[i].write(servoPosStart[i]);
     }
 }
 
 // Set the position of a servo
-void ServoControl::setServo(int idx, int angle) {
+void ServoControl::setServo(uint8_t idx, uint16_t angle) {
     if (this->servoPos[idx] != angle) { // Change only if the new angle is different
         angle = constrain(angle, this->servoPosMin[idx], this->servoPosMax[idx]);
         this->servos[idx].write(angle);
@@ -36,10 +36,10 @@ void ServoControl::setServo(int idx, int angle) {
 }
 
 // Set the position of all servos
-void ServoControl::setAllServos(int angles[]) {
-    int angle;
+void ServoControl::setAllServos(uint16_t angles[]) {
+    uint16_t angle;
 
-    for (int i=0; i<this->numberOfServos; i++)
+    for (uint8_t i=0; i<this->numberOfServos; i++)
         if (this->servoPos[i] != angles[i]) { // Change only if the new angle is different
             angle = constrain(angles[i], this->servoPosMin[i], this->servoPosMax[i]);
             this->servos[i].write(angle);
@@ -47,4 +47,4 @@ void ServoControl::setAllServos(int angles[]) {
         }
 }
 
-int ServoControl::getServo(int idx) { return this->servoPos[idx]; }
+uint16_t ServoControl::getServo(uint8_t idx) { return this->servoPos[idx]; }
